Store state time for provider nodes
In order to know how long a node has been in a state we'll store a state
timestamp similar to how it's currently done in Nodepool.

In addition we can use the state time to determin if a node has expired.
For that we'll remove the expiry time and just store the max-ready age
of the label. In combination with the state time we can then determine
if the node expired.

The max-ready age is stored as part of the node so that we don't need to
look up that piece of information by looping over providers and labels
every time we process a node.

Change-Id: I4bc74b8a837766e757368c7142a1a8090bc3a209