Use openstack endpoint for image format
Currently zuul-launcher is hardcoded to assume every cloud wants
a raw image.  We want to determine this value from the actual
cloud endpoint.  That is simple enough with openstack since it
is part of the cloud region configuration object that we pass to
the client connection constructor.

However, we use this value while parsing or deserializing the
provider config, and we don't currently have a mechanism to obtain
an endpoint without a complete provider.

To address that, this change adds an optional "extra" dictionary to
some zkobject deserialization methods so that in the code paths
we use to construct or deseralize a provider object, we can pass
in the connection registry and obtain a connection and endpoint
for use when parsing the config and creating OpenstackProviderImage
objects.

Once that is done, the openstack test is updated to use qcow2 in
the fake cloud to exercise the change.

Change-Id: I4ce75469c60e264c2786da15206b69f3dc020add