Improve launcher upload logging
This adds several log messages to indicate progress for a launcher
image upload.

It also updates the name of the uploaded image to include the
artifact uuid (so that we don't end up with a bunch of images
in the cloud named "debian-bullseye").

It also skips submitting the validate event if the image is already
validated.

Change-Id: Id11dd6d682c3d000d44e4e2367264cb028fc728f