WIP: Allow multiple errors for a single config item
For many (most? all? probably all) errors on configuration items,
we stop processing the item as soon as we encounter the first error.

That means that if a job references two projects, and both of them
do not exist, we would only store the error for the first missing
project and ignore the second one.

Zuul evaluates dynamic configuration changes to decide if they
introduce a new configuration error.  If a change introduces an error,
it is not permitted.  If, after parsing the config Zuul only sees
errors that it knows about already, it is permitted.

Together, these behaviors mean that if a configuration object has more
than one error, then in order to correct that, it must be done in
one change (or in one cycle of dependencies).  That is because Zuul
only knows about one of the errors.  If the user fixes the error that
Zuul knows about, then the other error will appear and Zuul will think
it's new (even though the underlying problem already existed).  If it
is impractical for a user to fix both errors simultaneously, the only
way out of this is to force-merge at least one of the changes.

To avoid this, we could avoid masking these errors.  This means that
we could accumulate as many errors about a single configuration object
as possible when parsing the config.  This would give users a better
view of all that is wrong with a configuration, at the cost of more
tenant configuration errors.

FOR CONSIDERATION:

* This change currently only implements unmasking a single type of error:
  the "job not found" error.  If we decide to go with this approach,
  we would repeat this pattern for other exceptions we raise.
* We can choose to implement this for only certain errors (like job/project
  not found) which are the most likely to be impacted by this sort of
  situation.  But implementing more than that probably won't hurt for
  the same reason: most masked errors are likely to be those.
* This could be beneficial to users because it may reduce the number of
  round trips, if, say, a new job definition has 3 things wrong with it.
* This could annoy users by greatly increasing the number of static
  tenant configuration errors.

Let's decide if this change or the status quo is preferable before
finishing implementation.

Change-Id: I1f8788056f7c7375ba7e91a30f70d98643eee4cd