Allow multiple errors for a single config item
For many (most? all? probably all) errors on configuration items,
we stop processing the item as soon as we encounter the first error.

That means that if a job references two projects, and both of them
do not exist, we would only store the error for the first missing
project and ignore the second one.

Zuul evaluates dynamic configuration changes to decide if they
introduce a new configuration error.  If a change introduces an error,
it is not permitted.  If, after parsing the config Zuul only sees
errors that it knows about already, it is permitted.

Together, these behaviors mean that if a configuration object has more
than one error, then in order to correct that, it must be done in
one change (or in one cycle of dependencies).  That is because Zuul
only knows about one of the errors.  If the user fixes the error that
Zuul knows about, then the other error will appear and Zuul will think
it's new (even though the underlying problem already existed).  If it
is impractical for a user to fix both errors simultaneously, the only
way out of this is to force-merge at least one of the changes.

To avoid this, we could avoid masking these errors.  This means that
we could accumulate as many errors about a single configuration object
as possible when parsing the config.  This would give users a better
view of all that is wrong with a configuration, at the cost of more
tenant configuration errors.

This change partially implements the above idea.  It allows multiple
configuration errors on a single object, as long as they originate in
validateReferences.  This is the most likely place for errors to
originate due to out-of-band zuul configuration changes (like project
renames).  Additionally, validateReferences failures do not cause the
object to be omitted from the layout (unlike failures in object parsing).
That simplifies this change since we don't need to take extra steps
to prevent that.

A later change may update object parsing to accumulate multiple errors
as well, but it will need to ensure the object is not added to the
layout.

Despite the additional config errors this will generate, this could
ultimately be beneficial to users because it may reduce the number of
round trips if a new job definition has multiple things wrong with it.

Change-Id: I1f8788056f7c7375ba7e91a30f70d98643eee4cd