Reorder filters on status page
Having the change filter as default (instead of the pipeline filter) was
requested by a lot of users. It also makes sense since the piplines are
more prominent now and empty pipelines are hidden by default. So it's
much less usable to search for pipelines as it was on the old status
page.

Therefore, change the ordering of the filter categories, so that the
change filter comes first and the pipeline filter comes last.

Change-Id: I50cc7153dd0691743c4ffcf23c486db7bda1e08a