Update include-vars to use the ref checkout on ref jobs
Users might well expect that if they use include-vars in a job that
runs on a ref (branch or tag), that Zuul would use the vars file in
that ref.  That was not the case, and instead Zuul would use either
an override checkout or the project default branch.  That is because
include-vars follows the same pattern as zuul roles.  Except that Zuul
roles have a different quirk: they will pick the playbook branch if
the role repo is the same as the playbook repo.  That smooths a lot
of edge cases with tags.  But that doesn't make sense for include-vars
which should not change depending on what playbook is running.

This means that it is nearly impossible to do something sensible with
include-vars on a ref job.  The only way to cause it to load the actual
values in the ref would be to set an override checkout to the ref itself
which is not practical for tag jobs.  This means most of the time tag
jobs with include-vars would end up using the project default branch
(master), which may be very different than what was expected if the
tag was made from some other branch.

To address this, the default behavior for include-vars is to use the
ref checkout if the include-vars project is the project of the triggering
ref.  This should produce the most intuitive behavior for users.  In
case the previous behavior of using floating values from master is
desirable, a new option, 'use-ref' can be set to 'false'.

Change-Id: Ifeb3b333fe6dbb021b2189d75ad710ae03c52371