Add most remaining openstack config options
This adds most of the remaining openstack configuration options;
at least it should be enough to launch a node in most clouds at
this point.  Tests are added (notably, the fip configurations
were not covered in nodepool).

The launcher is updated to output create and delete state machine
state changes to aid in debugging.

Change-Id: I6e79b2a09fd74ae867455ded52afc7c690ca4106