Openstack: don't try to delete a node with no external id
If our create state machine didn't get as far as saving an
external_id there is nothing for the delete state machine to do.

Nodepool would have checked this value before creating a delete
state machine, but in zuul we're considering external_id and similar
to be more private to the endpoint implementations, so we need
to check it there.

Change-Id: Ib9aa0b86cc7dd6bb14fd2347de8f2fa84f18ae49