Check zkobject size before writing
If we attempt to write a non-sharded zkobject that is larger than
1MiB, zk will drop our connection, and we will retry in an infinite
loop.  To avoid this, check the data size before writing and raise
an exception.  This will still likely cause errors in Zuul, but
it may limit the blast radius of the damage, and may be more
recoverable.

ZK has a 1MiB max data size for a znode but that includes some
other data like the key (it's really a packet buffer size).  To
avoid false negatives, we use the same value we use for sharding:
1MB, which gives 47KiB headroom.

This should be fine since we either expect objects to be much smaller
than 1MiB or sharded.  We're just trying to catch the cases where
we should have sharded something but didn't.

Change-Id: If9da8b3968a2db813895ca5aabb736da652d7f2b