Use colorblind-friendly color schema for Zuul UI
This adapts the colors used on the new status page to be better suitable
for colorblind people. As there are different types of colorblindness,
I've used a "colorblind simulator" to verifiy this color schema [1].

The new colors are basically the default Patternfly 4 colors with a
slightly different green tone, so it's easier to distinguish successful
(green) and failed (red) results.

To ensure easy access to those colors, I've introduced some reusable CSS
variables that are suitable for both, light and dark theme.

In addition to the colors, I've also removed the opacity used for the
square icons in the collapsed pipeline view on the status page to
increase the contrast between the different colors.

[1]: https://davidmathlogic.com/colorblind/#%235BA352-%23C9190B-%23F0AB00-%238A8D90-%232B9AF3

Change-Id: I4352dd9456a30a2306d298f675de49ece760faf4