Re-add volume attributes to aws driver
This adds the volume attributes which were originally omitted for
simplicity.

They add some extra complexity in that there are several ways they
may be relevant to a user:

1) A user may want to specify the volume attributes of a zuul image
   that will be created and uploaded to aws.
2) A user may want to override the volume attributes used in an
   existing cloud image.
3) A user may want to specify that when Zuul launches instances with
   a certain flavor, it should override the image's volume attributes
   with new values (example: a larger disk).
4) The same, but for a specific label only.

Therefore this change addresses each of those with the following:

1) The image stanza with type "zuul" in an aws provider can contain
   the volume attributes.  This specifies what will be in the uploaded
   AMI.
2) The image stanza with type "cloud" in an aws provider can contai
   the volume attributes.  This will override the value in the AMI.
3) The flavor stanza in an aws provider can contain the volume attrs.
4) Same for label.

There is a little ambiguity when used in the image stanza since in a
"zuul" image it is used for writing the AMI and when used with a "cloud"
image, it is used for overriding the AMI.  But the result in the same:
instances launched with that image will have those values.

The order of precedence is: labels override flavors override images.

On top of that, these values may also be set at the provider level in
which case they will be used as default values for all of the above.

To simplify and unify the new label/flavor/image fallback behavior,
we process labels and pull any values from flavors and images when
loading the config.  This lets any drivers that opt-in to this behavior
simply reference the final label object to get the attributes.

Change-Id: I029fbc1d00db46ceb12b76e2e0557c54660fa505