Visualize fragmentation of QueueItem progressbar
The QueueItem progressbar is basically a div that is made up of multiple
progress bars (one for each job). Since the progressbar has the same
width for each QueueItem, the width of each job's fragment in the
progressbar is based on the number of jobs that are part of the
QueueItem.

The current representation of the QueueItem progressbar comes with a few
drawbacks:

  1. It's not clearly visible how many jobs are part of the QueueItem.
  2. There might be multiple jobs with the same result (or running
     state) next to each other, but those can only be distinguished by
     hovering over the progress bar and waiting for the job name to
     appear via tooltip.
  3. Depending on the "color structure" of the progressbar there could
     be some weird color transitions.

By adding a small border to each job fragment, we can fix those issues
and provide a more appealing style that better fits into the overall
design of the new status page.

Since we don't apply the border to the first fragment, this fragment is
usually 2px wider than the others. However, this should only be an issue
if there are a lot of jobs (>50), so that the width of each fragments
shrinks to 1-3px. In "normal" cases this should not be visible.

Change-Id: If1585536bacff05b1a90f822f0cd98c19b39ab57