Allow waiting for tree cache ltime sync point
Most of the launcher API workes on cached data and uses operations that
are idempotent. This means that it's usually ok if the cache doesn't
have the latest data available yet.

However, in some cases (e.g. for min-ready) we need to know if the cache
has been sync up to a certain point in order to make further decisions
based on the data in the cache.

For this we'll create an ephemeral sync node per cache instance that is
updated on each sync. The tree cache will wait for the update event to
propagate via the persistent-recursive watch and then notify the waiter
via an event callback.

Change-Id: I09d89ff79269b62b7b50a71f852689b7e529fc55