zuul-web: Simplify tenant list
Loading the tenant list on systems with many tenants and a large number
of items can take a considerable amount of time. This is because
calculating the project and queue counts is quite heavy.

These counts on the tenant overview page might not be very useful and
are not displayed at all on the tenant dropdown list.

This change removes these counts from the tenant overview page table.
Also, it changes the `/tenants` endpoint so that it just delivers a list
of names of the currently configured tenants, which loads considerably
faster.

Change-Id: I13bc149fb6c08f1111745100a9b54580d1b4b38b