zuul-web: Simplify tenant list
Loading the tenant list on systems with many tenants and a large number
of items can take a considerable amount of time. This is because
calculating the project and queue counts is quite heavy.

These counts on the tenant overview page might not be very useful, and
are not displayed at all on the tenant dropdown list.

This change removes these counts from the tenant overview page table by
querying a new endpoint `/tenant-list`, that just delivers a list of
names of the currently configured tenants which loads considerably
faster. It retains the existing `/tenants` endpoint for compatibility
reasons (e.g. for third-party clients that might rely on it).

Change-Id: I13bc149fb6c08f1111745100a9b54580d1b4b38b