Update patch set 2
Patch Set 2: Code-Review-1

Thanks for this!  I have mixed feelings:

I think of the original version as not trying to provide information about specific jobs.  It's just trying to convey an overall view of:

  * Is this change mostly succeeding?
  * Is it mostly failing?
  * Is it mostly indeterminate?

In other words, if there's a whole lot of red, then something is seriously wrong with the change or the system.  If there's a lot of green and little red, then probably something small needs fixing.

If the user wants to know more information, expand the jobs listing.

You can think of it like a "sparkline" (a "sparkgraph?").  You're not supposed to be able to read actual values from a sparkline, you're supposed to see the overall trend very easily without focusing on the individual details (like numbers).  Is the trend going up?  Is it going down?  That's the level of detail we were trying to convey originally.

If we add more detail, then it may make it harder to quickly and intuitively get the overall view.

If people are having trouble picking out exactly the failing job with a tooltip, then one way to address that (which I think would be perfectly fine) is to remove the tooltip so the bar displays even less information.

If we decide that we do want to go with this approach, then I think we still may want to tweak this a bit.  I do think that having the wide (2px) dark bar between each job does make it harder to read the overall bar.  Compare these two:

https://imgur.com/C07imWs.png
https://imgur.com/oKiVifK.png

I find it's difficult to get the overall view because the individual divisions are distracting.  I'm counting bars instead of seeing a field of mostly green with a little red.  I didn't even notice the blue bars with the division; probably because they are closer to the dark border.

Perhaps we could address that by reducing the thickness of the bar (1px) instead?
And if that's not enough, maybe instead of being a border that separates, we could overlay a transparancy mask, so it just darkens the color that's already there?

Those are two ideas I have for trying to realize the benefits of showing the number of jobs without losing the "at-a-glance" benefit we have currently.

(BTW, this does illustrate your point about the initial one being too big.  I like to think that having a lot of jobs is pretty normal for Zuul, so even though we might not choose to fix that right away, I think we should still consider that to be a pretty normal case.  :)

Patch-set: 2
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1, 3c75b699c4f05f131f3311d90c0a72346598cd07