Clean up old status page resources
This removes the react components for the old status page that are not
used anywhere else. The following two components are kept because they
are still used for the single change panel:
  - pages/ChangeStatus
  - containers/status/ItemPanel

Although this panel is not directly linked from the new status page,
it's still accessible via the links that Zuul reports to the version
control system.

Additionally, clean up some CSS rules which are no longer used.

Change-Id: If6b60986551671eda9d53e88214b8354181176de