Use local cache sync instead of remote
Change I09d89ff79269b62b7b50a71f852689b7e529fc55 introduced a
zookeeper treecache sync point so that we could verify that the
node treecache was synced with the state in zk before starting
to launch min-ready nodes.  However, since we check min-ready
nodes each time we iterate through the main loop, this causes
a lot of ZK writes.  One option would be to try to reduce the
number of times we check min-ready (and we may yet do that),
but this patch takes a different approach.

This change removes the ZK sync point in favor of simply checking
that we have no outstanding events in our local treecache.  We
receive events from ZK very quickly, and if the system is slow,
we are still likely to have received events for the cache even
if we have not processed all of them.  The theory is that if we
ensure that the cache event queue has drained (to a certain point)
that should be nearly as good as waiting for the round trip to
zk as in the previous implementation.

We already accept that min-ready in a multi-launcher environment
may have some inaccuracies, so the idea here is that this
read-only best-effort approach should still be sufficient.

Change-Id: I9a485770300d4c1dacbf455bad3bce99e9eea0ff