Store job expansion in redux on status page
Currently the status page will forget the individual expansion
selections when navigating away and back, or also occasionally
when components re-render in such a way that a queue item component
is replaced in the DOM.

To make the page more stable and intuitive, this stores the expansion
state for jobs in redux.  It is keyed by the stable queue id, so that
even as queue items move around, the expansion will be remembered.

It will be remembered when navigating around the web ui, but reloading
the page will cause it to be forgotten.

Currently any change to the redux store will also trigger a complete
re-render of the status page due to the mapStateToProps function in
the PipelineSummary component.  To make this more efficient, that
is transformed to use the useSelector hook so that components are
not re-rendered unless the content actually changes.

Note that this is not a complete solution: a further change is needed
in order to remember the pipeline expansion.  Currently in order to
see this change in action after navigating away, the containing
pipeline must be manually expanded again.

Change-Id: Iab2626750d610af86c983f3463164e5034c6b375