Store queue expansion in redux on status page
This follows on the previous commit by storing the expansion state
of the queues in redux.  This means that queue expansion will be
remembered when navigating away from the status page and back.

The keys are the same for both the pipeline overview and summary,
so the same queues will be expanded in both views.  They are also
not keyed by tenant, and they are not cleared when changing tenants,
so the same queues would also be expanded when viewing different
tenants.  We might choose to change this behavior in the future.

Change-Id: If9f0d9d5e7a33858bfb4e04a762709d8b137ab0a