Clear status expansion memory on "expand all" toggle
Normally the redux memory of which queues and items are expanded
overrides the expand-all toggles (so that a user can expand or
collapse everything, then collapse or expand just the items of
interest).  But this is confusing if the user collapses an item
and then enables "expand all" -- not everything is expanded.

To resolve this, any time the user changes the expand all toggle,
reset the expansion state memory.

This applies to "expand all" on the overview page, and
"show all jobs" on the pipeline detail page.

Change-Id: Ic26f19e13eef9ac7830382d6b17a190b5c1ee0b3