Update patch set 5
Patch Set 5:

This does have a very small intentional memory leak.

We store the expansion state keyed by item uuid.  Once a user expands or collapses an item, we will store that key and a boolean on an object.  As the user expands or collapses more individual items, we will add more keys.

The only time this will be cleared is if the user reloads the page, or (in a later change in this stack), toggles the expand-all switch.

This patch does include a facility for removing the key entries, but as long as we want this memory to persist while navigating around to different pages in the app, it is difficult to use.  It's easy to clear the key on unmount, but from react's perspective, unmounting a component because we're navigating away looks the same as unmounting it because it was removed from the pipeline.

Therefore, the only reasonable way to only remove keys when items are removed from pipelines is to reconcile the items in the pipeline with the expansion keys each time the status json is updated.  That is somewhat computationally intensive given the deeply nested nature of that data structure, so for the moment, I have chosen to leave this very small memory leak in place.  If it every becomes a problem, performing the extra computation to clean it up is fairly straightforward.

Patch-set: 5