Cleanup unused blob store lock paths
The blob store does not automatically clean up its lock paths, and
now that we are using it more, the number of empty lock paths
is growing too large.

Correct this by deleting lock paths when they are no longer used.
Also, look for leaked lock paths in case they are missed.  This
will also correct Zuul installations with existing leaked paths.

However, if there are too many leaked lock paths, we may not be
able to list them without triggering a zookeeper error, so detect
that case and log a message instead.  This should only happen
on extremely large or active systems.  A script is included to help
cleanup those systems.

Change-Id: I7d2256e3ed5675c3a854c4ecdc7585ce8520b127