Delete blob store lock paths on unlock
The blob store is generally intended to be lightweight when it
comes to locks.  The only time a lock is needed is when putting
an item into the store, and that is because as a content-addressible
store, we need to protect against two actors putting the same
content at the same time.  Once put, the content never changes and
no locks are needed.

Meanwhile, the store itself is designed to be able to grow quite
large.  This discrepancy between the expected number of data entries
and the number of locks can be seen in the paths used for each:
the data paths have sub-paths with the first 2 characters of the
hash in order to make a deeper tree and avoid issues with the
number of zk child nodes.  The lock paths are not sub-divided in
this way.

To avoid a situation where the content store grows large but is
still limited by the number of lock paths, rather than keeping the
lock path around after the item is put, we will delete the lock
path immediately after releasing the lock.

If another actor has started locking, then we will ignore the
NotEmptyError caused by their lock node.  If the other actor has
not yet created a lock node, then they may encounter a NoNodeError
due to our deleting the lock path, but the locking routine used
is robust against that error and will retry.

Change-Id: If33284ca690a19bfffde145cd3a62622816ec70d