Report config error when zuul config is added
Issue: when there is a config error when the zuul config is added for the
first time, zuul would not report the config error.

Reason: in this case, the pipeline is empty for the project, and the
"Project not in pipeline" error would be reported (if configured), and
the real error (config error) is not reported.

Fix: check config error first for reporting, since config error is more
informative and by default would be reported.

Change-Id: I127d7c39657ad8ec88803c275c90e3769e8c7121