Batch fake build zookeeper additions
When skipping a job, we create a fake build entry in ZK and also
a database record.  When skipping a large number of jobs, we
create many such builds and after each one, we rewrite the entire
QueueItem object in ZooKeeper.  To avoid excessive unecessary
writes to ZK, create all the fake build objects and then add
them to the QueueItem at once, resulting in only a single ZK
write operation.

A later change will do the same for the database updates.

Also, remove an obsolete argument from the addBuild method.

Change-Id: I5852f2e260e3f961ac570fa1b90a27d2ac2959dc