Batch fake build database additions
When skipping a job, we create a fake build entry in ZK and also
a database record.  When skipping a large number of jobs, we
create many such builds and after each one, we rewrite the entire
QueueItem object in ZooKeeper.  To avoid excessive unecessary
database queries, create all the fake build objects and then add
them to the database in one transaction, with only one buildset
query.

Also, remove the "missing_buildset_okay" parameter and replace
it with an exception handler in order to localize the logic
around when it's okay for a buildset to be missing.

Change-Id: Ibfb553a7d4d408c2c870fb9e9a653ef35db21fd8