WIP: Add ZKObjectDict
To protect against accidentally mutating dictionary attributes that
are attached to ZKObjects, add a new ZKObjectDict class that ensures
that there is an active context before mutation.

Some instances where we mutated dictionaries outside of an active
context are updated to do so.  These are not bugs, however, because
they were immediately followed by an updateAttributes call which
would correctly write the result to ZK.

TODO:
* Consider doing the same for lists
* Consider explicit rather than implicit creation of these objects
  for performance purposes.

Change-Id: Ie97b0670db1bf4d1ed21502804cfc8ddedfc0980