Allow dependent jobs to force dependencies to run
If a two jobs (let's call them "parent" and "child") have file
matchers and child has a hard dependency on parent, if parent does
not match the files in the change, an exception will be raised
while freezing the job graph.

The current best mitigation for this is to ensure that the file
matches for such parent-child job pairs match in such a way that
parent will always run if child does.

This fix does not work if a user changes the definition of child
without changing the definition of parent.  In this case, the file
matchers for child will be disabled, but not parent, and the error
will be raised.  The only way to push through such a change would
be to also make a no-op change to parent.

This is inconvenient, and as long as we're thinking along those
lines, the ordinary situation with file matchers is also
inconevient.  Make both situations easier by ignoring the results
of file matchers for hard dependencies.  This behavior should
be more intuitive, and will simplify file matchers (in many cases,
it may now be possible to configure a job to effectively never
match a change, and it will then only run if there is a dependent
job that does run).

This does not apply to soft dependencies.

Change-Id: I7df2262881cb952b46b8a99ad5ccbe912c2e070b