Revert "Delete blob store lock paths on unlock"
This reverts commit 6768cf939724c6858ec98cfc45dd330fde488745.

In practice, there can be many actors (ie, schedulers) locking and
unlocking a blob near-simultaneously.  This patch added only one
retry which means it can really only handle a race condition with
one other actor.

Rather than fixing this with infinite retries, revert this patch
and rely on the existing use- and time-based blobstore lock cleanup.
That is slower but does have the advantage of not constantly deleting
and re-creating lock paths for busy nodes.

Change-Id: Idb56678f24f54a790505db03fc0e57f8aeb375e8