Support for mod_auth_openidc in front of Zuul web
We are protecting Zuul Web by mod_auth_openidc but the current "force
reload document behavior" does not handle the 401 response code with
data content returned by this module. Thus our users are faced to a
never loading page when navigating on the Zuul Web UI after the
session expired.

The current implementation supports 403 for mod_auth_mellon.

However mod_auth_openidc have an hardcoded behavior returning 401
and don't requires an extra request mechanism because it relies
on extra strategy like "the presence of a Sec-Fetch-Mode header
with a value that is not equal to "navigate"" to detect XHR queries [1].

This patch reloads the page in case of 401 response code from a server
other than identified as CherryPy.

[1] https://github.com/OpenIDC/mod_auth_openidc/blob/master/auth_openidc.conf#L939

Change-Id: I71e6215ebfef1bb52b8a3b0b04b6280712b60b93