Fix auth redirect problem when root url is accessed
Problem:
When api root auth is enabled, accessing the root url (without the
'/tenants' part) would prevent the auth callback page to be
redirected correctly, and hangs on the "Login in progress" page.

Reason:
In this case, the `zuul_auth_redirect` stores a empty string, which
in turn cause `user.redirect` to be empty, and will not perform the
redirection.

Fix:
Store `/` to `zuul_auth_redirect` as a default if it would
be empty.

Change-Id: Iab0fc5a8ae8931578c04a4de2eb9c1c332e9ca05