Remove implicit smart reconfig during startup
In the past (especially before the scale-out scheduler) the expectation
was that restarting a scheduler would also pick up changes to the local
tenant config.

In a multi-scheduler deployment this can become a problem when the
tenant config is updated on another scheduler while e.g. a rolling
update in Kubernetes is in progress.

Schedulers should use the tenant config that's stored in Zookeeper and
not rely on the local config to be up-to-date, which a smart-reconfig
implicitly assumes to be the case.

Change-Id: I01f22888a08075a9efc86c3ed978b20e2df2c6bf