Add an option to skip workspace checkouts
If users have very large repositories, it may take a considerable
amount of time to perform a git checkout of them on the executor,
and that checkout is unlikely to be used since it is the copy of
the repo on the workspace that will be used.

This change adds an option to skip the workspace checkout in that
case.  The zuul-jobs role prepare-workspace-git performs an
independent checkout of the zuul ref (from zuul variables) after
pushing the repo to the remote worker node and should work be
compatible with this with no changes.  The simpler and less-used
prepare-workspace role merely rsyncs and is therefore not compatible
with this change.  For this, and other reasons, the default is
still to check out all repos.

If the option is set to disable checkouts, we will also perform
sparse checkouts of the private repo copies that we use for
playbooks, roles, and include-vars.  This is to keep the spirit
of reducing time and space used for large repos while still
maintaining functionality.

If a repo is later used for line-comment mapping, we will perform
a checkout in the workspace after the job completes.

Repos where we perform speculative merges are checked out during
the merge process.  To keep consistent output regradless of whether
a repo has merged performed or not, we will clear the working trees
of repos where merges have been performed if we are skipping repo
checkouts.  This could produce a small amount of extra IO traffic
as files are deleted.

Change-Id: I22572d5f212a6e2c47d4e8ceaa51555a21ff5c08