Reduce empty space on status page
A lot of users complained about the amount of empty space on the new
status page. This change tackles the problem by reducing or removing
space in various ways, while still keeping the look-and-feel of the new
status page:

1. Remove the space between queues and pipelines and make the QueueItem
   cards more concise. This is mainly done by reducing or removing
   variuos paddings and reducing the size of the QueueItem progress bar.

2. Shorten the gap between the filter toolbar and the main content. I
   think we can improve the layout further and will do so in a follow-up
   change.

3. Reduce the font-size used for the pipeline name. Previously I made
   this larger to better differentiate the pipeline name from the queue
   name, but this made the whole page look like it was zoomed in to 150%
   by default (which wasn't the case).

4. Reduce the min-width of the gutter, so that more items (pipelines on
   the Status page and queue items on the PipelineDetails page) fit on
   the screen.

Finally, improve the CSS rules for the zuul-compact-card, so we end up
with less and more aligned rules that should be suitable for different
components.

Change-Id: Id455d2ba20a1e209aa61e1fa66bc436579834c28