Reduce empty space on status page
A lot of users complained about the amount of empty space on the new
status page. This change tackles the problem by reducing or removing
space in various ways, while still keeping the look-and-feel of the new
status page:

1. Remove the space between queues and pipelines and make the QueueItem
   cards more concise. This is mainly done by reducing or removing
   variuos paddings.

2. Shorten the gap between the filter toolbar and the main content. I
   think we can improve the layout further and will do so in a follow-up
   change.

3. Use plain cards (without border) to visualize pipelines. This reduces
   the number of nested borders when a queue or pipeline is expanded. It
   also gives the status page a more "flattened" look which I find quite
   nice. Finally, it allos to further reduce the paddings without the
   downside of squashing the card's content into the border.

Finally, improve the CSS rules for the zuul-compact-card, so we end up
with less and more aligned rules that should be suitable for different
components.

Change-Id: Id455d2ba20a1e209aa61e1fa66bc436579834c28