Further reduce workspace checkouts
A previous change added the option to skip workspace checkouts, but
the merger may still checkout those repos if it needs to merge a
change for a job.  That is because our merge operations rely on
the index and working tree to perform merges.  It should be possible
to perform merges by only changing the index.  We can therefore
configure git to perform a minimal (ie, top-level only) sparse checkout
of these repos.

This change does that.  We will still delete all contents of the
working tree afterwords to provide a consistent resulting working
tree state.  If we need to perform line-mapping, we will disable
the sparse checkout and perform a full checkout.

Change-Id: Iab3d7762f1fbfc9404d7abc086b43d8299e8b41f